extract tsconfig tests to new spec file; add failing test for "extends" from multiple configs
